+++ changelog.rst
@@ -3,6 +3,10 @@
 Change Log
 ##########
 
+- Unreleased
+
+  Add ``with-command-redo-break`` to end an active redo-chain.
+
 - Version 0.1 (2026-04-05)
 
   Initial release.